refactor(app): compose account section headers
