+++ packages/app/src/account/component/account-section-header/account-section-header.tsx
@@ -1,10 +1,9 @@
 import { AccountTypeEnum } from '@budgie/contracts';
 import { useLingui } from '@lingui/react/macro';
-import { Text, View } from 'react-native';
+import { Text } from 'react-native';
 
-import { useDisplayFormatDigits } from '../../../i18n/hook/use-display-format-digits.hook';
-import { useSettingsContext } from '../../../settings/context/settings.context';
 import { ACCOUNT_TYPE } from '../../constant/account-type.constant';
+import { AccountSectionHeaderFrame } from '../account-section-header-frame/account-section-header-frame';
 
 interface Props {
     readonly type: AccountTypeEnum;
@@ -13,15 +12,10 @@ interface Props {
 
 export const AccountSectionHeader = ({ type, total }: Props) => {
     const { t } = useLingui();
-    const { defaultInstrument } = useSettingsContext();
-    const formatDigits = useDisplayFormatDigits();
-
-    const formattedTotal = formatDigits(total, defaultInstrument.symbol);
 
     return (
-        <View className="bg-primary-reverse py-md -mx-5xl px-5xl flex-row justify-between items-center">
+        <AccountSectionHeaderFrame total={total}>
             <Text className="text-xs uppercase text-secondary-foreground">{t(ACCOUNT_TYPE[type])}</Text>
-            <Text className="text-xs text-secondary-foreground">{formattedTotal}</Text>
-        </View>
+        </AccountSectionHeaderFrame>
     );
 };
