+++ packages/app/src/account/component/bank-provider-section-header/bank-provider-section-header.tsx
@@ -5,9 +5,8 @@ import { Text, View } from 'react-native';
 import { isDefined } from '@rnw-community/shared';
 
 import { BankLogo } from '../../../@generic/component/bank-logo/bank-logo';
-import { useDisplayFormatDigits } from '../../../i18n/hook/use-display-format-digits.hook';
-import { useSettingsContext } from '../../../settings/context/settings.context';
 import { BANK_PROVIDER_TITLE } from '../../constant/bank-provider-title.constant';
+import { AccountSectionHeaderFrame } from '../account-section-header-frame/account-section-header-frame';
 
 interface Props {
     readonly provider: ExternalSourceEnum;
@@ -16,20 +15,16 @@ interface Props {
 
 export const BankProviderSectionHeader = ({ provider, total }: Props) => {
     const { t } = useLingui();
-    const { defaultInstrument } = useSettingsContext();
-    const formatDigits = useDisplayFormatDigits();
 
-    const formattedTotal = formatDigits(total, defaultInstrument.symbol);
     const titleDescriptor = BANK_PROVIDER_TITLE[provider];
     const title = isDefined(titleDescriptor) ? t(titleDescriptor) : provider;
 
     return (
-        <View className="bg-primary-reverse py-md -mx-5xl px-5xl flex-row justify-between items-center">
+        <AccountSectionHeaderFrame total={total}>
             <View className="flex-row items-center gap-sm">
                 <BankLogo bankProvider={provider} size={20} />
                 <Text className="text-xs uppercase text-secondary-foreground">{title}</Text>
             </View>
-            <Text className="text-xs text-secondary-foreground">{formattedTotal}</Text>
-        </View>
+        </AccountSectionHeaderFrame>
     );
 };
